+++ src/main/java/domain/amount/YearlyInstallmentInvestmentAmount.java
@@ -20,7 +20,7 @@ public class YearlyInstallmentInvestmentAmount implements InstallmentInvestmentA
 	}
 
 	@Override
-	public int calMonthlyInterest(InterestRate interestRate) {
+	public double calMonthlyInterest(InterestRate interestRate) {
 		return interestRate.calMonthlyInterest(amount);
 	}
 }
